Rename logging to avoid import errors
